Reorganise crate poms to use pallet-crates as parent
This solves an issue where the maven release plugin would get the wrong
directory for the git repository, and fail when tagging.

It also introduces the site phase to generate marginalia and autodoc
ouput.  This seems to fail for autodoc as it doesn't seem to run inside
a phase.

Remove push of commits during release:prepare.